feat: load children async
