--- .vscode/extensions.json
@@ -1,9 +1,0 @@
-{
-  "recommendations": [
-    "nrwl.angular-console",
-    "esbenp.prettier-vscode",
-    "firsttris.vscode-jest-runner",
-    "dbaeumer.vscode-eslint",
-    "redhat.vscode-yaml"
-  ]
-}
