+++ .vscode/settings.json
@@ -29,5 +29,7 @@
       "apps/api-gateway/router.yaml",
       "apps/api-gateway/router.prod.yaml"
     ]
-  }
+  },
+  "typescript.preferences.importModuleSpecifier": "project-relative",
+  "javascript.preferences.importModuleSpecifier": "project-relative"
 }
